Move and improve ContactPlayerCallback
ContactPlayerCallback should be in Player.hpp for now since it is
specific to the player. In the future we may want to have it in
Entity.hpp and apply to Entity in general, so that all entities are
affected by triggers.
